Extract spec setup into support file
